メイン処理とクラス定義のファイルを分けた
